+++ tests/services/scanner/scanner_dao_test.py
@@ -17,7 +17,7 @@
 import ast
 import unittest
 
-from google.cloud.forseti.scanner import dao as scanner_dao
+from google.cloud.forseti.services.scanner import dao as scanner_dao
 from tests.services.utils.db import create_test_engine
 from tests.unittest_utils import ForsetiTestCase
